+++ CHANGELOG.md
@@ -29,6 +29,7 @@ contents to a new release section (e.g. _[0.x]_).
 - Add reciprocal power rule i.e. `[1/a^n] = a^-n` and collect integers with rational exponents (PLUT-625)
 - Make sure leading factor of 1 is not removed when followed by division sign and relax rule for turning divisions into
   fractions (PLUT-664)
+- Fix parsing the variable 'd' (PLUT-671)
 
 ### Changed